complete invoices

+++ app/Http/Controllers/Panel/ProductController.php
@@ -119,6 +119,10 @@ class ProductController extends Controller
     {
         $this->authorize('products-delete');
 
+        if ($product->invoices()->exists()){
+            return response('این محصول در پیش فاکتور هایی موجود است',500);
+        }
+
         $product->delete();
         return back();
     }
